generating the new model from the original model
